requirements.txt
elasticsearch==7.13.4
fake-headers==1.0.2
google-cloud-secret-manager==2.16.1
google-cloud-storage==2.9.0
python-dotenv==1.0.0
smart-open==6.3.0
tqdm==4.65.0
